Always use the Sqlite database to store session data
Previously we would store this data in SQLite, and a memory based storage that express provides.  This created a scenario where a token could be dleeted from the Sqlite database, but not from the express memory store.  As a result the UI could render, make a query and then have to reload to get a new token.  With this change the UI should never load if there is no token.  Instead, it will request a token before rendering the UI